fix: i2c: change buffer due to CPU endianness before/after memcpy
Issue found when VBUS of USB3.0 didn't work in BE kernel.
Found that memcpy copy the bytes from u8 array to u32 start from high byte
{ 0xa, 0xb, 0xc, 0xd}, after memcpy to u32:
BE: a0b0c0d
LE: d0c0b0a

The issue that the order of the messages is the same in BE, and LE.
As wrokaroud after/before memcpy change the order of the write/read
according to the CPU endianness.

Change-Id: Id43e5c0935a275d33d266f2a5053019b80f4a846
Signed-off-by: Hanna Hawa <hannah@marvell.com>