task_isolation: arch/tile: enable task isolation functionality
We add the necessary call to task_isolation_start() in the
prepare_exit_to_usermode() routine.  We already unconditionally
call into this routine if TIF_NOHZ is set, since that's where
we do the user_enter() call.

We add calls to task_isolation_interrupt() in places where exceptions
may not generate signals to the application.

Change-Id: Ibeea838b17ab2c4b68b7e58262e03fa25ee0942c
Signed-off-by: Chris Metcalf <cmetcalf@mellanox.com>
Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>
Signed-off-by: Alex Belits <abelits@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/9862
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Stefan Chulski <Stefan.Chulski@cavium.com>