net: phy: sfp: handle non-wired SFP connectors
SFP connectors can be solder on a board without having any of their pins
(LOS, i2c...) wired. In such cases the SFP link state cannot be guessed,
and the overall link status reporting is left to other layers.

In order to achieve this, a new SFP_DEV status is added, named UNKNOWN.
This mode is set when it is not possible for the SFP code to get the
link status and as a result the link status is reported to be always UP
from the SFP point of view.

Change-Id: I29aaf6a9d34d69e4d4ff7dbf61aea5f614ef8473
Signed-off-by: Antoine Tenart <antoine.tenart@bootlin.com>