dts: armada-37xx: reserve 16M memory for tee
These memory are used for TEE code, data, communication channel
and shared memory with Non-Trusted world.

For the TEE code and data section, it will be configured
unacceable to Non-Trusted World by memory controller.

The rest memories are either used for communication or shared
memory between Trusted-World and Non-trusted World, which is
TEE implementation defined and may be managed by the TEE driver.

Change-Id: I19536da77513e39a195d5fffeeeb0068c089f7e5
Signed-off-by: wangwen <wangwen@marvell.com>