dts: a8040-db: include 8040-db.dtsi instead of 8040.dtsi
We should use the following structure in the device-tree:
        armada-8040.dtsi - SoC common definitions
        armada-8040-db.dtsi - Board common definitions.
        armada-8040-db-X.dts - Specific board definitions

Currently, in armada-8040-db.dts, we include armada-8040.dtsi
directly, skipping armada-8040-db.dtsi which creates duplications
in the device-tree.

This patch fixes this issue, buy replacing the include in
armada-8040-db.dts.

The following definitions are added with this change:
- spi-flash@0
- several sdhci@780000 definitions.
- cp0_vccq_sd0
- several NAND definitions

Cleanup of the duplication will be done in a separate patch.

Change-Id: I0fefbbea9f91186ba5265e755f1e2a9565deb945
Signed-off-by: Igal Liberman <igall@marvell.com>