net: mvneta: add comphy initialization
On Armada 37xx platforms the common phy (comphy) must be configured to
work in appropriate mode. Hitherto the mventa driver was relying on
firmware/bootloader configuration. Since the comphy driver for a3700 is
now supported in Linux, the mvneta driver can take advantage and
initialize appropriate comphy to relevant mode.

Change-Id: I3e29346b9fed26388f95e144787f74d03edb6533
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>