phy: marvell: comphy: refactor driver to ease adding new driver variant
This change does not change any functionality of existing driver. The
aim of this rework is to prepare current driver for new variant of
comphy driver found in Armada 37xx SoC.

This patch introduces the comphy data which abstracts differences
between current and upcoming variant of comphy driver. It specifies the
SMC format, available modes, number of lanes and ports which are
different in current and upcoming variant.

Additionally in upcoming variant of comphy driver the comphy address space
is divided into 4 ranges, therefore requesting resources is now dependent
on number of declared resources.

Change-Id: If516b50a0c1e25042ddfdcf42ce6c10bc2f3f830
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>