cpufreq: add support for CPU DVFS based on SCMI message protocol
On some ARM based systems, a separate Cortex-M based System Control
Processor(SCP) provides the overall power, clock, reset and system
control including CPU DVFS. SCMI Message Protocol is used to
communicate with the SCP.

This patch adds a cpufreq driver for such systems using SCMI interface
to drive CPU DVFS.

All changes for SCMI is taken from 4.18 kernel but some part of it is
dependent on features that is available only in 4.18. For a particular
function, implementation is done based on 4.14 infrasturcture, thus this
patch was brought in from previous version of 4.14 based SCMI patches
to support SCMI on 4.14.

Cc: linux-pm@vger.kernel.org
Acked-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>
Signed-off-by: Sujeet Baranwal <sbaranwal@caviumnetworks.com>