mvpp2: add MUSDK compatibility for RSS
In prior mvpp2 implementation, enum mvpp2_cls_lkp_type contained also
enum for QoS (vlan and dscp). MUSDK still expects these enums so
MVPP2_CLS_LKP_DEFAULT should have the value of 3 and not 1.
Once QoS will be introduced in mvpp2, MUSDK will be reviewed and
adjusted accordingly to the new implementation, but meanwhile this
is the minimum change needed for RSS to work in MUSDK.

Change-Id: I6e630ead8c756edbacd7c5bb75f24f280688054d
Signed-off-by: Alan Winkowski <walan@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/4743
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Stefan Chulski <Stefan.Chulski@cavium.com>