net: mvpp2: fix in parser sram shift function
This patch fixes an issue when setting the parser sram shift
information. In case the shift was negative, the shift sign
bit was set but then was overridden when setting the shift value

This fix should solve issues seen in IPv6, since the negative
shift is only used for IPv6 entries.

Change-Id: Ia56b7610b09fa09fae7bac1443bf17b8dc23e839
Signed-off-by: Alan Winkowski <walan@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/1568
Tested-by: sa_ip-sw-jenkins
Tested-by: Stefan Chulski <Stefan.Chulski@cavium.com>
Reviewed-by: Stefan Chulski <Stefan.Chulski@cavium.com>