octeontx: assign and free PKIND for loop interface
There is no need for static binding. We may allocate as well from the
remaining free PKINDs (instead of widening the pool already assigned
for other LBK based interfaces).

Loop interface can be distinguished because its port ID at the point
of domain creation is -1. Thanks to that we can recognize it and assign
PKIND from pool destined for loop interfaces.

Change-Id: I23fd2da75c3e74eb162928616c7efc2cbee55124
Signed-off-by: Angela Czubak <aczubak@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/14800
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Stanislaw Kardach <Stanislaw.Kardach@cavium.com>
Reviewed-by: Vidhya Vidhyaraman <Vidhya.Raman@cavium.com>