scmi: increase timeout for polling
At times, the time taken by SCP to service the request is more
than 100uS and that results into transactions not being completed.

This manifests into various failures, one of them being the sysfs
reads of hwmon-sensors values.

Increasing the timeout value allows more time to SCP to respond
to the request correctly.

Bug: IPBUSW-2393

Change-Id: Ib30eee75e6182eb2746f8ec90f6fc3a561289afe
Signed-off-by: Sujeet Baranwal <sbaranwal@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/16420
Reviewed-by: Sujeet Kumar Baranwal <Sujeet.Baranwal@cavium.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>