scmi: increase timeout for polling
At times, the time taken by SCP to service the request is more
than 100uS and that results into transactions not being completed.

This manifests into various failures, one of them being the rclk
frequencies not changing because of insufficient timeout.

Increasing the timeout value allows more time to SCP to respond
to the request correctly.

Bug: IPBUSW-2537

Change-Id: I235baa04118c049d0444e3ab5eae7b73c7a8c3e5
Signed-off-by: Sujeet Baranwal <sbaranwal@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/16712
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>