Revert "Revert "drivers: pci: Add support for multiple PCI MSI""
This reverts commit 12f33cdf205c5919b688fbbf4b197926134ff2e8.

Restore multiple PCI MSI to un-break PCIe support for ipq6018.

This is a workaround. The right fix is to update DT for GIC based MSI
support. But we are missing needed hardware information to fix this
properly in DT as commit e910db2590 ("arm64: dts: ipq5018: enable GIC
based MSI support") does for ipq5018.